requirements.txt
google-cloud-speech
google-cloud-translate
google-cloud-texttospeech
pyaudio
sounddevice
numpy
    